do not remove slashes in file path (#53) (#54)
do not remove slashes in file path (#53) (#54)

On Windows hosts stripping (back)slashes causes weird file paths in the
UI. A leading (forward) slash is also not appropriate if these paths are
absolute with leading drive letter.

Do not remove slashes and omit the fixed leading one to fix the output.